update: order trucking dan order delete cek jurnal

+++ app/Http/Controllers/OrderController.php
@@ -325,9 +325,13 @@ class OrderController extends Controller
 
     public function destroy(Order $order)
     {
+        $cekJurnalTakTerhapus = Jurnal::where('order_id', $order->id)->whereNull('deleted_at')->count();
+        if ($cekJurnalTerhapus > 0) {
+            return back()->with('danger','Data job ini tidak bisa dihapus karena biaya sudah tercatat!');
+        }
         $order->delete();
         HutangPelayaran::where('order_id',$order->id)->delete();
-        return back()->with('success','Data berhasil dihapus');
+         return back()->with('success','Data berhasil dihapus');
     }
 
     public function copy(Order $order)
